requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
diskcache==5.6.3
filelock==3.16.1
fsspec==2024.10.0
gguf==0.10.0
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
llama_cpp_python==0.3.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
protobuf==4.25.5
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
sympy==1.13.3
tokenizers==0.21.0
torch==2.2.2+cpu
tqdm==4.67.1
transformers==4.47.0
typing_extensions==4.12.2
urllib3==2.2.3
